requirements/base.txt
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.4
argon2-cffi==23.1.0 # https://github.com/hynek/argon2_cffi
argon2-cffi-bindings==21.2.0
cffi==1.16.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
whitenoise==6.6.0

# Documentacion
# drf-yasg==1.21.7
drf-spectacular==0.27.2
setuptools==69.5.1

# DJANGO
# ------------------------------------------------------------------------
dj-database-url==2.2.0
Django==4.2.16 # https://www.djangoproject.com/
django-filter==24.3
django-model-utils==5.0.0 # https://github.com/jazzband/django-model-utils
django-crispy-forms==2.3 # https://django-crispy-forms.readthedocs.io/en/latest/install.html
crispy-bootstrap4==2024.1
# Django REST Framework
djangorestframework==3.14.0 # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.3.1 # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/
django-cors-headers==4.4.0 # https://github.com/adamchainz/django-cors-headers

# WEASYPRINT CONFIGURATION
Brotli==1.1.0
cssselect2==0.7.0
fonttools==4.47.2
html5lib==1.1
pillow==10.2.0
pydyf==0.10.0
pyphen==0.14.0
six==1.16.0
tinycss2==1.3.0
weasyprint==62.0
webencodings==0.5.1
zopfli==0.2.3

# AWS

boto3==1.35.25
django-storages==1.14.4

#CELERY && REDIS
celery==5.4.0 # https://github.com/celery/celery
redis==5.0.8 # https://github.com/andymccurdy/redis-py
django-celery-results==2.5.1
